Move time formatting to utils folder
